Crawl/requirements.txt
requests
pandas
mysql-connector-python
sqlalchemy
bs4
